[seller] refactor(173): 이미지 삭제 및 모달관련 hook으로 분리
