+++ apps/seller/src/features/products/create/create-form-thumbnail-upload/use-product-thumnail-form.hook.ts
@@ -1,3 +1,5 @@
+import { useState } from 'react'
+
 import { useFormContext } from 'react-hook-form'
 
 import { CreateProductForm } from '../create-form/product-create.types'
@@ -7,6 +9,8 @@ export function useProductThumbnailForm() {
 
   const mainImage = form.watch('mainImage')
   const extraImages = form.watch('extraImages') || []
+
+  const [deleteTarget, setDeleteTarget] = useState<number | 'main' | null>(null)
 
   // 필수 입력 사항 판별: 대표 이미지가 있으면 true
   const isFormField = mainImage !== null
@@ -19,12 +23,41 @@ export function useProductThumbnailForm() {
     form.setValue('extraImages', files, { shouldValidate: true })
   }
 
+  const handleFileChange = (
+    e: React.ChangeEvent<HTMLInputElement>,
+    type: 'main' | 'extra',
+  ) => {
+    const files = e.target.files
+    if (!files) return
+
+    if (type === 'main') {
+      handleMainImageChange(files[0])
+    } else {
+      const remainingSlots = 9 - extraImages.length
+      const newFiles = Array.from(files).slice(0, remainingSlots)
+      handleExtraImagesChange([...extraImages, ...newFiles])
+    }
+    e.target.value = ''
+  }
+
+  const handleImageDelete = () => {
+    if (deleteTarget === 'main') {
+      handleMainImageChange(null)
+    } else if (typeof deleteTarget === 'number') {
+      const newImages = extraImages.filter((_, i) => i !== deleteTarget)
+      handleExtraImagesChange(newImages)
+    }
+    setDeleteTarget(null)
+  }
+
   return {
     form,
     mainImage,
     extraImages,
     isFormField,
-    handleMainImageChange,
-    handleExtraImagesChange,
+    deleteTarget,
+    setDeleteTarget,
+    handleFileChange,
+    handleImageDelete,
   }
 }
